Makefile: fix kernel.fit usage
The "more m5 splitting commit removed the kernel.fit target replacing it
with kernel_m5.fit. Unfortunately this brakes the rtk target as the
kernel.fit can no longer be found.

This patch updates parts of the Makefile to fix the problem however I'm
not replacing all kernel.fit accesses as I'm not sure if nor and nor_ipl
targets should use the m5 or breadbee kernel image.