Makefile: some more kernel.fit dependency fixes
kernel.fit target does not exist anymore. I've tried to change all
places where the Makefile is using it by something sane - mostly using
kernel_breadbee.fit instead.

For "nor" target I've removed it as the current commands seems not be
using it at all. For "copy_kernel_to_sd", I've added kernel_m5.fit as a
dependency as the commands are using the product of this target. I've
also made sure that kernel_breadbee.fit is properly handled by clean and
is on a list of phony targets.